+++ src/Control/Settings/Attributes/AdditionalAttributesPosition.php
@@ -36,7 +36,7 @@ class AdditionalAttributesPosition extends GlobalSetting
      */
     public function fieldOptions(): Field
     {
-        return \FormSchema\Generator\Field::input($this->inputName())->inputType('text')->getSchema();
+        return \FormSchema\Generator\Field::textInput($this->inputName());
     }
 
     /**