+++ tests/Completion/CompletionConditions/EventFired/EventFiredTest.php
@@ -6,6 +6,7 @@ use BristolSU\Support\ActivityInstance\ActivityInstance;
 use BristolSU\Support\Completion\CompletionConditions\EventFired\EventFired;
 use BristolSU\Support\Events\Contracts\EventRepository;
 use BristolSU\Support\Tests\TestCase;
+use FormSchema\Fields\SelectField;
 
 class EventFiredTest extends TestCase
 {
@@ -29,15 +30,20 @@ class EventFiredTest extends TestCase
 
         $condition = new EventFired('module1', $eventRepository->reveal());
 
+        $groups = $condition->options()->groups();
+        $this->assertCount(1, $groups);
 
+        $fields = $groups[0]->fields();
+        $this->assertCount(1, $fields);
 
-        $this->assertEquals(1, count($condition->options()->fields()));
-        $this->assertEquals('event_type', $condition->options()->fields()[0]->model());
-        $this->assertEquals('select', $condition->options()->fields()[0]->type());
+        $field = $fields[0];
+
+        $this->assertInstanceOf(SelectField::class, $field);
+        $this->assertEquals('event_type', $field->getId());
         $this->assertEquals([
-            ['id' => 'event1', 'name' => 'name1'],
-            ['id' => 'event2', 'name' => 'name2'],
-        ], $condition->options()->fields()[0]->values());
+            ['id' => 'event1', 'value' => 'name1'],
+            ['id' => 'event2', 'value' => 'name2'],
+        ], $field->getSelectOptions());
     }
 
     /** @test */