+++ tests/Settings/Definition/UserSettingTest.php
@@ -144,7 +144,7 @@ class UserSettingTestDummyUserSetting extends UserSetting
      */
     public function fieldOptions(): Field
     {
-        return \FormSchema\Generator\Field::input($this->inputName())->inputType('text')->getSchema();
+        return \FormSchema\Generator\Field::textInput($this->inputName());
     }
 
     /**